+++ account-service/src/main/kotlin/exp/yaremchuken/account_service/repository/AccountCredentialsRepository.kt
@@ -3,4 +3,4 @@ package exp.yaremchuken.account_service.repository
 import exp.yaremchuken.account_service.repository.entity.AccountCredentials
 import org.springframework.data.repository.CrudRepository
 
-interface AccountCredentialsRepository: CrudRepository<AccountCredentials, Int>
+interface AccountCredentialsRepository: CrudRepository<AccountCredentials, Long>
