+++ account-service/src/main/kotlin/exp/yaremchuken/account_service/repository/AccountRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AccountRepository: CrudRepository<Account, Int>
+interface AccountRepository: CrudRepository<Account, Long>
