+++ account-service/src/main/kotlin/exp/yaremchuken/account_service/repository/AccountSettingsRepository.kt
@@ -3,4 +3,4 @@ package exp.yaremchuken.account_service.repository
 import exp.yaremchuken.account_service.repository.entity.AccountSettings
 import org.springframework.data.repository.CrudRepository
 
-interface AccountSettingsRepository: CrudRepository<AccountSettings, Int>
+interface AccountSettingsRepository: CrudRepository<AccountSettings, Long>
